Update password.rb
We are getting bots/spiders attempting to visit bogus password reset URLs for users that do not have a reset_token_digest created on their Identity record.  This will result in SCrypt throwing an exception because the authenticate method will attempt to pass it a nil token.  Prevent this by checking the value before sending it to SCrypt.